Inline `Table` into `AssetsTable`
